Reorganized database files to be contained in a single package.
