+++ backend/database/assignments/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
-from assignments import models
+
+from database.assignments import models
+
 # Register your models here.
 admin.site.register(models.Student)
 admin.site.register(models.FlaggedStudent)
