+++ backend/database/courses/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from courses import models
+from database.courses import models
 
 # Register your models here.
 admin.site.register(models.Professor)
