+++ backend/database/courses/models.py
@@ -2,7 +2,7 @@
 Courses Database Models.
 """
 from django.db import models
-from users import models as user_model
+from database.users import models as user_model
 from django.conf import settings
 
 class Professor(models.Model):
@@ -18,7 +18,7 @@ class Class(models.Model):
 
     # Explicitly reference the Professor model from the users app
     professors = models.ManyToManyField(
-        "courses.Professor",
+        "database.courses.Professor",
         #the join table will happen at the ProfessorClassSection instead of being created automatic bewteen prof and class
         through="ProfessorClassSection",
     )
@@ -35,7 +35,7 @@ class Semester(models.Model):
 
 class ProfessorClassSection(models.Model):
     """Mapping Professors to Class Sections."""
-    professor = models.ForeignKey("courses.Professor", on_delete=models.CASCADE, related_name="profclassect")
+    professor = models.ForeignKey("database.courses.Professor", on_delete=models.CASCADE, related_name="profclassect")
     class_instance = models.ForeignKey(Class, on_delete=models.CASCADE)
     semester = models.ForeignKey(Semester, on_delete=models.CASCADE)
     section_number = models.IntegerField(blank=True, null=True)
