+++ backend/database/courses/urls.py
@@ -3,7 +3,7 @@ Courses URLs.
 """
 from rest_framework.routers import DefaultRouter
 from rest_framework_nested.routers import NestedDefaultRouter
-from courses.views import (ProfessorVS, ProfessorClassSectionVS, SemesterVS, ClassVS)
+from database.courses.views import (ProfessorVS, ProfessorClassSectionVS, SemesterVS, ClassVS)
 from django.urls import path, include
 
 class_router = DefaultRouter()
