+++ backend/database/users/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from users import models
+from database.users import models
 
 # Register your models here.
 admin.site.register(models.User)
